requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
channels==4.3.1
redis==6.4.0
django-filter==25.1
numpy==2.3.2
Pillow==10.4.0